fix some warnings and notes

+++ Marshmallows/UIAlertViewDelegate.h
@@ -9,7 +9,7 @@
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
 
-typedef void (^UIAlertViewCallback)(long buttonClicked, BOOL canceled);
+typedef void (^UIAlertViewCallback)(NSInteger buttonClicked, BOOL canceled);
 
 @interface UIAlertViewDelegate : NSObject <UIAlertViewDelegate>
 {
